Refactor EventPipeEventSource
I extracted the FastSerializer deserialization logic out of EventPipeEventSource and put it into a separate FastSerializerObjectParser object. When using file format versions <= 5 this parser uses Deserializer to read all the objects. In V6+ we'll use a different (simpler) parser to iterate through blocks. I also added the V6 header detection logic but if it detects the V6 currently the code throws NotImplException.

Added a V6 format parsing test case which triggers the NotImplException